ticket service setup
